created multiple galleries that only users can see and use, only let image owners to add/delete images

+++ assignment/3/app/app.js
@@ -1,3 +1,4 @@
+var crypto = require('crypto');
 var path = require('path');
 var express = require('express');
 var app = express();
@@ -5,43 +6,173 @@ var app = express();
 var multer = require('multer');
 var upload = multer({dest: 'uploads/'});
 
+var cookieParser = require('cookie-parser');
+app.use(cookieParser());
+
 var Datastore = require('nedb');
 var images = new Datastore({ filename: 'db/images.db', autoload: true, timestampData : true});
 var comments = new Datastore({ filename: 'db/comments.db', autoload: true, timestampData : true});
+var users = new Datastore({ filename: 'db/users.db', autoload: true, timestampData: true});
 
 var fs = require('fs');
+var session = require('express-session');
+app.use(session({
+    secret: 'keyboard cat',
+    resave: false,
+    saveUninitialized: true
+}));
 
 var bodyParser = require('body-parser');
 app.use(bodyParser.json());
+
+var privateKey = fs.readFileSync( 'server.key' );
+var certificate = fs.readFileSync( 'server.crt' );
+var config = {
+        key: privateKey,
+        cert: certificate
+};
+var https = require('https');
 
 app.use(express.static('frontend'));
 
 
 app.use(function (req, res, next){
-    console.log("HTTP request", req.method, req.url, req.body);
+    console.log("HTTPS request", req.method, req.url, req.body);
     return next();
 });
 
 
+//Authentication
+
+var checkPassword = function(user, password){
+        var hash = crypto.createHmac('sha512', user.salt);
+        hash.update(password);
+        var value = hash.digest('base64');
+        return (user.saltedHash === value);
+};
+
+
+// signout, signin
+
+app.get('/api/signout/', function (req, res, next) {
+    req.session.destroy(function(err) {
+        if (err) return res.status(500).end(err);
+        return res.end();
+    });
+	res.clearCookie('username');
+});
+
+app.post('/api/signin/', function (req, res, next) {
+    if (!req.body.username || !req.body.password)
+    	return res.status(400).send("Bad Request");
+
+    //look for the user in the db and confirm that the password is right
+    users.findOne({username: req.body.username}, function(err, user){
+        if (err) return res.status(500).end(err);
+        if (!user || !checkPassword(user, req.body.password))
+        	return res.status(401).end("Unauthorized");
+        req.session.user = user;
+        res.cookie('username', user.username);
+        user.curr_user = user.username;
+        res.json(user);
+        return next();
+    });
+});
+
+
+
+
+
+
+
+
+
+
+//req.cookies returns the set of cookies
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 //Create
 
-//adds the image info to the image database, and if picture is a file upload it
-app.post('/api/images/', upload.single('picture'), function(req, res, next){
+//creates a new user
+app.post('/api/users/', function(req, res, next){
+
+	//check if the user already exists
+	users.findOne({username: req.body.username}, function(err, user){
+		if(err) return res.status(500).end(err);
+		if(user) return res.status(409).end("Username " + req.body.username + " already exists");
+
+		//create the salted hash
+	    var salt = crypto.randomBytes(16).toString('base64');
+	    var hash = crypto.createHmac('sha512', salt);
+	    hash.update(req.body.password);
+	    
+	    //link the new user to the previous one
+	    users.findOne({right: null}, function(err, result){
+	    	if(err) return res.status(500).end(err);
+	    	var left = null;
+	    	if(result)
+	    		left = result.username;
+
+		    //insert the new user
+		    users.insert({username: req.body.username, salt: salt, saltedHash: hash.digest('base64'), left: left, right: null}, function(err, user){
+		    	if(err) return res.status(500).end(err);
+
+		    	//link the new user to the previous one
+		    	if(result){
+		    		result.right = user.username;
+		    		users.update({username: result.username}, result);
+		    	}
+		    	req.session.user = user;
+		    	res.cookie('username', user.username);
+		    	user.curr_user = user.username;
+		    	return res.json(user);
+		    });
+	    });
+	});
+});
+
+
+
+
+//user req.cookies.username to figure out what user is currently making calls to api
+
+
+
+
+//adds the image info to the gallery's image database, and if picture is a file upload it
+app.post('/api/gallery/:gallery/images/', upload.single('picture'), function(req, res, next){
+	if((!req.cookies) || (req.cookies.username !== req.params.gallery))
+		return res.status(401).end("Forbidden");
 
 	var file = req.file;
 	if(file === undefined)
 		file = req.body.picture;
 
-	//link the previous image to the new one
-	images.findOne({right: null}, function(err, result){
+	//link the gallery's previous image to the new one
+	images.findOne({gallery: req.params.gallery, right: null}, function(err, result){
+		if(err) return res.status(500).end(err);
 		var left = null;
 		if(result)
 			left = result._id;
 
-		//add the image info to the images database
-		images.insert({picture: file, title: req.body.title, author: req.body.author, left: left, right: null}, function(err, insert_result){
+		//add the image info to the gallery's images database
+		images.insert({gallery: req.params.gallery, picture: file, title: req.body.title, author: req.cookies.username, left: left, right: null}, function(err, insert_result){
+			if(err) return res.status(500).end(err);
 
-			//link the new image to the previous one
+			//link the new image to the previous one in the gallery
 			if(result){
 				result.right = insert_result._id;
 				images.update({_id: result._id}, result);
@@ -52,23 +183,28 @@ app.post('/api/images/', upload.single('picture'), function(req, res, next){
 	});
 });
 
-//adds the comment to the comments database and links it to the image with id imageId
-app.post('/api/images/:imageId/comments/', function(req, res, next){
-	var image_id = req.params.imageId;
+//adds the comment to the comments database and links it to the image with id imageId in gallery
+app.post('/api/gallery/:gallery/images/:imageId/comments/', function(req, res, next){
+	if((!req.cookies) || (req.cookies.username === undefined))
+		return res.status(401).end("Forbidden");
 
-	//link the previous comment with the same image id to the new comment 
-	comments.findOne({newer_comment: null, image_id: image_id}, function(err, result){
+	//link the previous comment with the same imageId in gallery to the new comment 
+	comments.findOne({gallery: req.params.gallery, newer_comment: null, image_id: req.params.imageId}, function(err, result){
+		if(err) return res.status(500).end(err);
 		var older_comment = null;
 		if(result)
 			older_comment = result._id;
 
 		//add the comment to the comments database
-		comments.insert({image_id: image_id, author: req.body.author, message: req.body.message, date: req.body.date, older_comment: older_comment, newer_comment: null}, function(err, insert_result){
+		comments.insert({gallery: req.params.gallery, image_id: req.params.imageId, author: req.cookies.username, message: req.body.message, date: req.body.date, older_comment: older_comment, newer_comment: null}, function(err, insert_result){
+			if(err) return res.status(500).end(err);
 
 			//link the new comment to the previous comment
 			if(result){
 				result.newer_comment = insert_result._id;
-				comments.update({_id: result._id}, result);
+				comments.update({_id: result._id}, result, function(err){
+					if(err) return res.status(500).end(err);
+				});
 			}
 			res.json(insert_result._id);
 			return next();
@@ -79,30 +215,62 @@ app.post('/api/images/:imageId/comments/', function(req, res, next){
 
 //Read
 
+//gets the users gallery or returns nothing if there is no signed in user
+app.get('/api/gallery/', function(req, res, next){
+	if((!req.cookies) || (req.cookies.username === undefined))
+		return;
+
+	users.findOne({username: req.cookies.username}, function(err, user){
+		user.curr_user = req.cookies.username;
+		res.json(user);
+		return next();
+	});
+});
+
+//returns the gallery if the supplied gallery name exists
+app.get('/api/gallery/:gallery/', function(req, res, next){
+	console.log(req.cookies)
+	console.log(!req.cookies)
+	console.log(req.cookies.username === undefined)
+	if((!req.cookies) || (req.cookies.username === undefined))
+		return res.status(401).end("Forbidden");
+
+	users.findOne({username: req.params.gallery}, function(err, gallery){
+		gallery.curr_user = req.cookies.username;
+		res.json(gallery);
+		return next();
+	});
+});
+
 //stops 404 favicon errors from http://stackoverflow.com/questions/35408729/express-js-prevent-get-favicon-ico
 app.get('/favicon.ico', function(req, res, next) {
     res.sendStatus(204);
     return next();
 });
 
-//gets the image data for the image with the given id
-app.get('/api/images/:id/', function(req, res, next){
-	var id = req.params.id;
-	if(id !== "first"){
+//gets the image data for the image with the given id in the given gallery
+app.get('/api/gallery/:gallery/images/:id/', function(req, res, next){
+	if((!req.cookies) || (req.cookies.username === undefined))
+		return res.status(401).end("Forbidden");
 
-		//finds the image with _id id
-		images.findOne({_id: id}, function(err, result){
-			if(result)
+	if(req.params.id !== "first"){
+
+		//finds the image with the given id in the given gallery
+		images.findOne({_id: req.params.id, gallery: req.params.gallery}, function(err, result){
+			if(err) return res.status(500).end(err);
+			if(result){
+				result.curr_user = req.cookies.username;
 				res.json(result);
-			else
-				res.status(404).json("Image with id " + id + " does not exist");
+			}
+			else return res.status(404).json("Image with id " + req.params.id + " does not exist in gallery " + req.params.gallery);
 			return next();
 		});
 	}
 	else{
 
-		//gets the first image
-		images.findOne({}).sort({createdAt: 1}).exec(function(err, result){
+		//gets the first image in the gallery
+		images.findOne({gallery: req.params.gallery}).sort({createdAt: 1}).exec(function(err, result){
+			if(err) return res.status(500).end(err);
 			res.json(result);
 			return next();
 		});
@@ -110,11 +278,13 @@ app.get('/api/images/:id/', function(req, res, next){
 });
 
 //gets the picture file with the given id
-app.get('/api/images/:id/picture/', function(req, res, next){
-	var id = req.params.id;
+app.get('/api/gallery/:gallery/images/:id/picture/', function(req, res, next){
+	if((!req.cookies) || (req.cookies.username === undefined))
+		return res.status(401).end("Forbidden");
 
 	//return the image url
-	images.findOne({_id: id}, function(err, result){
+	images.findOne({_id: req.params.id, gallery: req.params.gallery}, function(err, result){
+		if(err) return res.status(500).end(err);
 		if(result){
 			if(typeof(result.picture) == "string")
 				res.json(result.picture);
@@ -123,15 +293,17 @@ app.get('/api/images/:id/picture/', function(req, res, next){
 		        res.sendFile(path.join(__dirname, result.picture.path));
 	    	}
     	}
-    	else
-            res.status(404).json("Image with id " + req.params.id + " does not exists");
+    	else return res.status(404).json("Image with id " + req.params.id + " does not exists in gallery " + res.params.gallery);
     	return next();
 	});
 });
 
-//gets comments for the image with 'imageId', starting at 'firstComment' where the comments are sorted as specified and stopping after limit is reached
-//the default for limit is 10 and the default for sort is decreasing
-app.get('/api/images/:imageId/comments/:firstComment/', function(req, res, next){
+//gets comments for the image with 'imageId' in the given gallery, starting at 'firstComment' where the comments are sorted as specified and stopping after
+//limit is reached, the default for limit is 10 and the default for sort is decreasing
+app.get('/api/gallery/:gallery/images/:imageId/comments/:firstComment/', function(req, res, next){
+	if((!req.cookies) || (req.cookies.username === undefined))
+		return res.status(401).end("Forbidden");
+
 	var firstComment = req.params.firstComment;
 	var imageId = req.params.imageId;
 	var limit = req.query.limit;
@@ -140,26 +312,23 @@ app.get('/api/images/:imageId/comments/:firstComment/', function(req, res, next)
 	//make sure that if limit or sort was provided that they are valid
 	if(limit === undefined)
 		limit = 10;
-	else if(isNaN(limit)){
-		res.status(400).json("Invalid arguments. Limit must be a number and " + limit + " is not");
-		return next();
-	}
+	else if(isNaN(limit))
+		return res.status(400).json("Invalid arguments. Limit must be a number and " + limit + " is not");
 	if(sort === undefined)
 		sort = "decreasing";
-	else if((sort !== "decreasing") && (sort !== "increasing")){
-		res.status(400).json("Invalid arguments. Sort must be a decreasing or increasing and " + sort + " is not");
-		return next();
-	}
+	else if((sort !== "decreasing") && (sort !== "increasing"))
+		return res.status(400).json("Invalid arguments. Sort must be a decreasing or increasing and " + sort + " is not");
 
-	var search = {_id: firstComment};
+	var search = {_id: firstComment, gallery: req.params.gallery};
 	if(firstComment === "last")
-		search = {image_id: imageId};
+		search = {image_id: imageId, gallery: req.params.gallery};
 
-	//get the current comment
+	//get the current comment for the image in the current gallery
 	comments.findOne(search).sort({createdAt: -1}).exec(function(err, comment){
+		if(err) return res.status(500).end(err);
 
-		//default database query that will get the comments in decreasing order of creation
-		var query = {image_id: imageId};
+		//default database query that will get the comments in decreasing order of creation for the given image in the given gallery
+		var query = {image_id: imageId, gallery: req.params.gallery};
 		var order = {createdAt: -1};
 
 		if(comment){
@@ -177,6 +346,7 @@ app.get('/api/images/:imageId/comments/:firstComment/', function(req, res, next)
 
 		//get comments resulting from the query, stopping after limit is reached
 		comments.find(query).sort(order).limit(limit).exec(function(err, result){
+			if(err) return res.status(500).end(err);
 			if((sort == "increasing"))
 				res.json(result.reverse());
 			else
@@ -194,21 +364,24 @@ app.get('/api/images/:imageId/comments/:firstComment/', function(req, res, next)
 
 //Delete
 
-//deletes the image with the given id
-app.delete('/api/images/:id/', function(req, res, next){
-	var id = req.params.id;
+//deletes the image with the given id from the given gallery
+app.delete('/api/gallery/:gallery/images/:id/', function(req, res, next){
+	if((!req.cookies) || (req.cookies.username !== req.params.gallery))
+		return res.status(401).end("Forbidden");
 
-	//get the image to be deleted
-	images.findOne({_id: id}, function(err, result){
+	//get the image to be deleted from the given gallery
+	images.findOne({_id: req.params.id, gallery: req.params.gallery}, function(err, result){
+		if(err) return res.status(500).end(err);
 		if(result){
 
-			//find the next image's id
+			//find the next image's id in the gallery
 			var next_id = result.left;
 			if(next_id === null)
 				next_id = result.right;
 
 			//update the left image's right image link
 			images.findOne({_id: result.left}, function(err, left_res){
+				if(err) return res.status(500).end(err);
 				if(left_res){
 					left_res.right = result.right;
 					images.update({_id: result.left}, left_res);	
@@ -217,32 +390,42 @@ app.delete('/api/images/:id/', function(req, res, next){
 
 			//update the right image's left image link
 			images.findOne({_id: result.right}, function(err, right_res){
+				if(err) return res.status(500).end(err);
 				if(right_res){
 					right_res.left = result.left;
 					images.update({_id: result.right}, right_res);
 				}
 			});
 
-			images.remove({_id: id});
-			fs.unlink(result.picture.path);
-			res.json(next_id);
+			images.remove({_id: result._id}, function(err){
+				if(err) return res.status(500).end(err);
+				if(result.picture.path)
+					fs.unlink(result.picture.path);
+				res.json(next_id);
+				return next();
+			});
 		}
-		else
-            res.status(404).json("Image " + id + " does not exists");		
-		return next();
+		else return res.status(404).json("Image " + req.params.id + " does not exists in the gallery " + req.params.gallery);
 	});
 });
 
 //deletes the comment with the given id
-app.delete('/api/comments/:id/', function(req, res, next){
-	var id = req.params.id;
+app.delete('/api/gallery/:gallery/images/:imageId/comments/:id/', function(req, res, next){
+	if((!req.cookies) || (req.cookies.username !== req.params.gallery))
+		return res.status(401).end("Forbidden");
 
-	//get the comment to be deleted
-	comments.findOne({_id: id}, function(err, comment){
+	//get the comment to be deleted from the given image in the given gallery
+	comments.findOne({_id: req.params.id, image_id: imageId, gallery: req.params.gallery}, function(err, comment){
+		if(err) return res.status(500).end(err);
 		if(comment){
+
+			//make sure that only the comment author or gallery owner can delete a comment
+			if((comment.author !== req.cookies.username) && (comment.gallery !== req.cookies.username))
+				return res.status(401).end("Forbidden");
 
 			//update the older_comment's newer_comment link
 			comments.findOne({_id: comment.older_comment}, function(err, older_comment){
+				if(err) return res.status(500).end(err);
 				if(older_comment){
 						older_comment.newer_comment = comment.newer_comment;
 						comments.update({_id: older_comment._id}, older_comment);
@@ -251,27 +434,29 @@ app.delete('/api/comments/:id/', function(req, res, next){
 
 			//update the newer_comment's older_comment link
 			comments.findOne({_id: comment.newer_comment}, function(err, newer_comment){
+				if(err) return res.status(500).end(err);
 				if(newer_comment){
 					newer_comment.older_comment = comment.older_comment;
 					comments.update({_id: newer_comment._id}, newer_comment);
 				}
 			});
 
-			comments.remove({_id: comment._id});
-			res.json(id);
+			comments.remove({_id: comment._id}, function(err){
+				if(err) return res.status(500).end(err);
+			});
+			res.json(req.params.id);
 			return next();
 		}
-		else
-            res.status(404).json("Comment with id " + id + " does not exists");
+		else return res.status(404).json("Comment with id " + req.params.id + " does not exists in gallery " + req.params.gallery);
 	});
 });
 
 
 app.use(function (req, res, next){
-    console.log("HTTP Response", res.statusCode);
+    console.log("HTTPS Response", res.statusCode);
 });
 
 
-app.listen(3000, function () {
+https.createServer(config, app).listen(3000, function () {
   console.log('App listening on port 3000');
 });
